// src/app/types/climates.ts
export enum EPlanetClimates {
  FRIGID = "frigid",
  TROPICAL = "tropical",
  FROZEN = "frozen",
  MOIST = "moist",
  POLLUTED = "polluted",
  UNKNOWN = "unknown",
  SUBARTIC = "subartic",
  TEMPERATE = "temperate",
  ARTIFICIAL_TEMPERATE = "artificial temperate",
  MURKY = "murky",
  HUMID = "humid",
  ARTIC = "artic",
  SUPERHEATED = "superheated",
  HOT = "hot",
  ROCKY = "rocky",
  ARID = "arid",
  WINDY = "windy",
}